Update product y customer
