+++ src/app/app.routes.ts
@@ -7,11 +7,11 @@ import { CategoryComponent } from './modules/product/component/category/category
 import { ProductComponent } from './modules/product/component/product/product.component';
 import { authenticationGuard } from './modules/auth/authentication.guard';
 import { HomeComponent } from './modules/home/home.component';
-import { ProductDetailsComponent } from './modules/product/component/product-details/product-details.component';
+import { ProductImageComponent } from './modules/product/component/product-image/product-image.component';
 import { CustomerComponent } from './modules/customer/component/customer/customer.component';
 import { RegionComponent } from './modules/customer/component/region/region.component';
 import { InvoiceComponent } from './modules/invoice/component/invoice/invoice.component';
-import { CustomerDetailsComponent } from './modules/customer/component/customer-details/customer-details.component';
+import { CustomerImageComponent } from './modules/customer/component/customer-image/customer-image.component';
 import { ProductByCategoryComponent } from './modules/product/component/product-by-category/product-by-category.component';
 
 export const routes: Routes = [
@@ -35,7 +35,7 @@ export const routes: Routes = [
     },
     {
         path: 'product/:gtin',
-        component: ProductDetailsComponent
+        component: ProductImageComponent
     },
     {
         path: 'register',
@@ -58,7 +58,7 @@ export const routes: Routes = [
 
     {
         path: 'customer/:rfc',
-        component: CustomerDetailsComponent
+        component: CustomerImageComponent
     },
