+++ src/app/modules/customer/_dto/dto-customer-list.ts
@@ -1,8 +1,16 @@
 export class DtoCustomerList {
 
-    customer_id: number = 0;
-    name: string = "";
-    surname: string = "";
-    rfc: string = "";
-    status: number = 0;
+    customer_id: number;
+    name: string;
+    surname: string;
+    rfc: string;
+    status: number;
+
+    constructor() {
+        this.customer_id = 0;
+        this.name = '';
+        this.surname = '';
+        this.rfc = '';
+        this.status = 0;
+    }
 }
