+++ src/app/modules/customer/_model/customer-image.ts
@@ -1,0 +1,10 @@
+export class CustomerImage {
+
+    customer_image_id: number;
+    image: string;
+
+    constructor() {
+        this.customer_image_id = 0;
+        this.image = '';
+    }
+}
