--- src/app/modules/customer/_model/customer/customer-image.ts
@@ -1,5 +1,0 @@
-export class CustomerImage {
-    
-    customer_image_id: number = 0;
-    image: string = "";
-}
