+++ src/app/modules/customer/_model/region.ts
@@ -1,0 +1,13 @@
+export class Region {
+    region_id: number;
+    region: string;
+    tag: string;
+    status: number;
+
+    constructor() {
+        this.region_id = 0;
+        this.region = "";
+        this.tag = "";
+        this.status = 0;
+    }
+}
