--- src/app/modules/customer/_model/region/region.ts
@@ -1,7 +1,0 @@
-export class Region {
-    
-    region_id: number = 0;
-    region: string = "";
-    tag: string = "";
-    status: number = 0;
-}
