+++ src/app/modules/customer/component/customer-image/customer-image.component.css
