+++ src/app/modules/invoice/_dto/dto-cart-details.ts
@@ -1,8 +1,8 @@
-import { Product } from "../../product/_model/product/product";
+import { Product } from "../../product/_model/product";
 
 
 export class DtoCartDetails {
-    
+
     cart_id: number = 0;
     rfc: string = "";
     gtin: string = "";
