+++ src/app/modules/product/_dto/dto-product-list.ts
@@ -1,9 +1,17 @@
 export class DtoProductList {
-    
-    product_id: number = 0;
-    product: string = "";
-    gtin: string = "";
-    price: number = 0;
-    category_id: number = 0;
-    status: number = 0;
+    product_id: number;
+    product: string;
+    gtin: string;
+    price: number;
+    category_id: number;
+    status: number;
+
+    constructor() {
+        this.product_id = 0;
+        this.product = '';
+        this.gtin = '';
+        this.price = 0;
+        this.category_id = 0;
+        this.status = 0;
+    }
 }
