+++ src/app/modules/product/_model/category.ts
@@ -1,13 +1,14 @@
 export class Category {
-    category_id: number = 0;
-    category: string = "";
-    tag: string = "";
-    status: number = 0;
 
-    constructor(category_id: number, category: string, tag: string, status: number) {
-        this.category_id = category_id;
-        this.category = category;
-        this.tag = tag;
-        this.status = status;
-    }
+  category_id: number;
+  category: string;
+  tag: string;
+  status: number;
+
+  constructor() {
+    this.category_id = 0;
+    this.category = '';
+    this.tag = '';
+    this.status = 0;
+  }
 }
