--- src/app/modules/product/_model/category/category.ts
@@ -1,7 +1,0 @@
-export class Category {
-
-  category_id: number = 0;
-  category: string = "";
-  tag: string = "";
-  status: number = 0;
-}
