+++ src/app/modules/product/_model/product-image.ts
@@ -1,5 +1,12 @@
 export class ProductImage {
-    product_id: number = 0;
-    product_image_id: number = 0;
-    image: string = "";
+
+    product_image_id: number;
+    product_id: number;
+    image: string;
+
+    constructor() {
+        this.product_image_id = 0;
+        this.product_id = 0;
+        this.image = '';
+    }
 }
