--- src/app/modules/product/_model/product/product-image.ts
@@ -1,6 +1,0 @@
-export class ProductImage {
-    
-    product_image_id: number = 0;
-    product_id: number = 0;
-    image: string = "";
-}
