--- src/app/modules/product/_model/product/product.ts
@@ -1,11 +1,0 @@
-export class Product {
-    
-    product_id: number = 0;
-    product: string = "";
-    gtin: string = "";
-    description: string = "";
-    price: number = 0;
-    stock: number = 0;
-    category_id: number = 0;
-    status: number = 0;
-}
