+++ src/app/modules/product/_service/product-image.service.ts
@@ -2,7 +2,7 @@
 import { HttpClient, HttpResponse } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
-import { ProductImage } from '../_model/product/product-image';
+import { ProductImage } from '../_model/product-image';
 import { api_dwb_uri } from '../../../shared/api-dwb-uri';
 import { ApiResponse } from '../../../shared/api-response';
 
@@ -25,7 +25,7 @@ export class ProductImageService {
   getProductImages(product_id: number): Observable<HttpResponse<ProductImage[]>> {
     return this.http.get<ProductImage[]>(api_dwb_uri + this.source + "/" + product_id, { observe: 'response' });
   }
-  
+
   deleteProductImage(product_image_id: number): Observable<HttpResponse<ApiResponse>> {
     return this.http.delete<ApiResponse>(api_dwb_uri + this.source + "/" + product_image_id, { observe: 'response' });
   }
